+++ src/types/ReadingStatus.ts
@@ -1,0 +1,1 @@
+export type ReadingStatus = 'unread' | 'reading' | 'done';
